Rename navBar to NavBar file
